Rename SelectAddressModalGroup directory
