+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/AddDeviceForm.vue
