+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/DeviceConnectingModal.vue
