+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/SelectDeviceForm.vue
