+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/__test__/AddAddressForm.spec.js
