+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/__test__/SelectAddressForm.spec.js
