+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/__test__/SelectNetworkAddressModal.spec.js
