+++ kolibri/core/assets/src/views/sync/SelectDeviceModalGroup/api.js
