+++ kolibri/plugins/user_profile/assets/src/views/ChangeFacility/SelectFacility.vue
@@ -79,9 +79,9 @@
   import client from 'kolibri.client';
   import urls from 'kolibri.urls';
   import BottomAppBar from 'kolibri.coreVue.components.BottomAppBar';
-  import useSavedAddresses from '../../../../../../core/assets/src/views/sync/SelectAddressModalGroup/useSavedAddresses.js';
-  import useDynamicAddresses from '../../../../../../core/assets/src/views/sync/SelectAddressModalGroup/useDynamicAddresses.js';
-  import AddDeviceForm from '../../../../../../core/assets/src/views/sync/SelectAddressModalGroup/AddDeviceForm';
+  import useSavedAddresses from '../../../../../../core/assets/src/views/sync/SelectDeviceModalGroup/useSavedAddresses.js';
+  import useDynamicAddresses from '../../../../../../core/assets/src/views/sync/SelectDeviceModalGroup/useDynamicAddresses.js';
+  import AddDeviceForm from '../../../../../../core/assets/src/views/sync/SelectDeviceModalGroup/AddDeviceForm';
   import useMinimumKolibriVersion from '../../../../../../core/assets/src/composables/useMinimumKolibriVersion';
 
   export default {
